# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is a Educational Application project",
        "title": "EdTech API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "https://ilessons.cloud/go/api",
    "basePath": "/",
    "paths": {
...
        "/lessons": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Find all lessons",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Lesson"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Create Lesson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson_title",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lesson_description",
                        "name": "description",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Lesson_subject_id",
                        "name": "subject_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Topic_order",
                        "name": "order",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_level",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_interest",
                        "name": "interest",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "minimum age for viewing lesson",
                        "name": "target_age_min",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "maximum age for viewing lesson",
                        "name": "target_age_max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video",
                        "name": "video_data",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video_filename",
                        "name": "video_filename",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video_MIME-type",
                        "name": "video_mime_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Lessons_video_duration_seconds",
                        "name": "duration_sec",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "lessons_video_is_published",
                        "name": "is_published",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    }
                }
            }
        },
        "/lessons/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "find lesson by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Lesson"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "Update Lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lesson_title",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lesson_description",
                        "name": "description",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Lesson_subject_id",
                        "name": "subject_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Topic_order",
                        "name": "order",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_level",
                        "name": "level",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_interest",
                        "name": "interest",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "minimum age for viewing lesson",
                        "name": "target_age_min",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "maximum age for viewing lesson",
                        "name": "target_age_max",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video",
                        "name": "video_data",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video_filename",
                        "name": "video_filename",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lessons_video_MIME-type",
                        "name": "video_mime_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Lessons_video_duration_seconds",
                        "name": "duration_sec",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "lessons_video_is_published",
                        "name": "is_published",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "lessons"
                ],
                "summary": "delete lesson by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid Payload",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.Lesson": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_sec": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "interest": {
                    "type": "string"
                },
                "is_published": {
                    "type": "boolean"
                },
                "level": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "subject_id": {
                    "description": "under question",
                    "type": "integer"
                },
                "target_age_max": {
                    "type": "integer"
                },
                "target_age_min": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "video_filename": {
                    "type": "string"
                },
                "video_mime_type": {
                    "type": "string"
                }
            }
        },
...
    },
...
}